Ensure buffers are really flushed
Following 5a7145925a000f6f5bfdcd2ca83fe54c68685170 , this commit
ensures that all buffers are emptied when FlushOutput, by adding a
return value to that function, that indicates the length of the buffer
after the flush function was called.

While the current code does this by retrying until the flushing is
complete, we might want to refactor this in the future, allowing for
scenarios where flushing just isn't possible, instead of being 'stuck'
in an infinite loop.